Fix: Update file structure #252
