+++ apps/web/frontendnew/src/components/Button/Custom.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+const Custom = ({ href, text, buttonClass, arrowClass }) => {
+  return (
+    <a href={href} className={buttonClass}>
+      {text}
+      <span className={arrowClass} aria-hidden="true">
+        &rarr;
+      </span>
+    </a>
+  );
+};
+
+export default Custom;
