+++ apps/web/frontendnew/src/pages/Home/Home.js
@@ -1,44 +1,14 @@
 import { useContext } from "react";
-import {
-  BoltIcon,
-  DevicePhoneMobileIcon,
-  GlobeAltIcon,
-  ScaleIcon,
-} from "@heroicons/react/24/outline";
 import { Helmet } from "react-helmet";
 import { ThemeContext } from "../../contexts/ThemeModeProvider";
+import MainContainer from "../../components/Container/MainContainer";
+import Features from "../../components/Container/Features";
+import Testimonials from "../../components/Container/Testimonials";
 
 const Home = () => {
   const { theme } = useContext(ThemeContext);
 
   const dark = theme === "dark";
-
-  const features = [
-    {
-      name: "Cloud Synced",
-      description:
-        "Tired of not having the right info handy when you need it? Noteslify automatically saves notes online and syncs them to all your devices.",
-      icon: GlobeAltIcon,
-    },
-    {
-      name: "Reminders",
-      description:
-        "Create and assign your task with due dates and reminders. Find notes by searching for keywords, titles, or tags.",
-      icon: ScaleIcon,
-    },
-    {
-      name: "Capture Infinitely",
-      description:
-        "Store as many notes you can safely backed up on cloud. Get all arranged in one place safely and securely.",
-      icon: BoltIcon,
-    },
-    {
-      name: "Web application",
-      description:
-        "Many note-taking apps lack a fully functional web application. Noteslify offers a complete lineup of features from any major browser.",
-      icon: DevicePhoneMobileIcon,
-    },
-  ];
 
   return (
     <>
@@ -49,271 +19,14 @@ const Home = () => {
           content="Your Privacy Friendly, Open Source. Alternative to EverNote."
         />
       </Helmet>
-      <>
-        {/* ### MAIN START */}
-        <main>
-          {/* ### HERO START */}
-          <div className="relative px-6 lg:px-8">
-            <div className="mx-auto max-w-3xl pt-20 pb-32 sm:pt-48 sm:pb-40">
-              <div className="">
-                <div className="hidden sm:mb-8 sm:flex sm:justify-center">
-                  <div
-                    data-theme={theme}
-                    className=" announce relative overflow-hidden rounded-full py-1.5 px-4 text-sm leading-6 ring-1 ring-gray-900/10 hover:ring-gray-900/20"
-                  >
-                    <span data-theme={theme} className="color text-gray-600">
-                      Announcing our next round of funding.{" "}
-                      <a href="/" className="font-semibold text-blue-600">
-                        <span className="absolute inset-0" aria-hidden="true" />
-                        Read more <span aria-hidden="true">&rarr;</span>
-                      </a>
-                    </span>
-                  </div>
-                </div>
-                <div>
-                  <h1 className="text-4xl font-bold tracking-tight sm:text-center sm:text-6xl">
-                    Your Privacy Friendly, Open Source. Alternative to EverNote.
-                  </h1>
-                  <p
-                    className={`mt-6 text-lg leading-8 ${
-                      dark ? "text-gray-400" : "text-gray-700"
-                    } sm:text-center`}
-                  >
-                    Capture your notes, files, and life’s work all in one secure
-                    place.
-                  </p>
-                  <div className="mt-8 flex gap-x-4 sm:justify-center">
-                    <a
-                      href="/"
-                      className="inline-block rounded-lg bg-blue-600 px-4 py-1.5 text-base font-semibold leading-7 text-white shadow-sm ring-1 ring-blue-600 hover:bg-blue-700 hover:ring-blue-700 transition-all"
-                    >
-                      Get started for free
-                      <span className="text-blue-200" aria-hidden="true">
-                        &rarr;
-                      </span>
-                    </a>
-                    <a
-                      href="/"
-                      className={`inline-block rounded-lg px-4 py-1.5 text-base font-semibold leading-7 text-gray-900 ring-1 ring-gray-900/10 hover:ring-gray-900/20 ${
-                        dark &&
-                        "text-gray-200 ring-1 ring-gray-300/10 hover:ring-gray-900/20"
-                      } transition-all`}
-                    >
-                      Visit Github
-                      <span className="text-gray-400" aria-hidden="true">
-                        &rarr;
-                      </span>
-                    </a>
-                  </div>
-                </div>
-                <div className="absolute inset-x-0 top-[calc(100%-13rem)] -z-10 transform-gpu overflow-hidden blur-3xl sm:top-[calc(100%-30rem)]">
-                  <svg
-                    className="relative left-[calc(50%+3rem)] h-[21.1875rem] max-w-none -translate-x-1/2 sm:left-[calc(50%+36rem)] sm:h-[42.375rem]"
-                    viewBox="0 0 1155 678"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      fill="url(#ecb5b0c9-546c-4772-8c71-4d3f06d544bc)"
-                      fillOpacity=".3"
-                      d="M317.219 518.975L203.852 678 0 438.341l317.219 80.634 204.172-286.402c1.307 132.337 45.083 346.658 209.733 145.248C936.936 126.058 882.053-94.234 1031.02 41.331c119.18 108.451 130.68 295.337 121.53 375.223L855 299l21.173 362.054-558.954-142.079z"
-                    />
-                    <defs>
-                      <linearGradient
-                        id="ecb5b0c9-546c-4772-8c71-4d3f06d544bc"
-                        x1="1155.49"
-                        x2="-78.208"
-                        y1=".177"
-                        y2="474.645"
-                        gradientUnits="userSpaceOnUse"
-                      >
-                        <stop stopColor="#9089FC" />
-                        <stop offset={1} stopColor="#FF80B5" />
-                      </linearGradient>
-                    </defs>
-                  </svg>
-                </div>
-              </div>
-            </div>
-          </div>{" "}
-          {/* ### HERO END */}
-        </main>
-      </>
-      {/* ### FEATURES START */}
-      <div
-        className={`${
-          dark && "bg-[#030414]"
-        } bg-white py-24 sm:py-32 lg:py-40 transition-all`}
-        id="features"
-      >
-        <div className="mx-auto max-w-7xl px-6 lg:px-8">
-          <div className="sm:text-center">
-            <h2 className="text-lg font-semibold leading-8 text-blue-600">
-              Features
-            </h2>
-            <p
-              className={`${
-                dark && "text-gray-200"
-              } mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl`}
-            >
-              Why choose Noteslify?
-            </p>
-            <p
-              className={`${
-                dark && "text-gray-400"
-              } mx-auto mt-6 max-w-2xl text-lg leading-8 text-gray-600`}
-            >
-              Make plans. Keep a journal. Clip anything from the web. Noteslify
-              lets you collect information quickly, organize it easily, and find
-              it when you need it.
-            </p>
-          </div>
-
-          <div className="mt-20 max-w-lg sm:mx-auto md:max-w-none">
-            <div className="grid grid-cols-1 gap-y-16 md:grid-cols-2 md:gap-x-12 md:gap-y-16">
-              {features.map((feature) => (
-                <div
-                  key={feature.name}
-                  className="relative flex flex-col gap-6 sm:flex-row md:flex-col lg:flex-row"
-                >
-                  <div className="flex h-12 w-12 items-center justify-center rounded-xl bg-blue-500 text-white sm:shrink-0">
-                    <feature.icon className="h-8 w-8" aria-hidden="true" />
-                  </div>
-                  <div className="sm:min-w-0 sm:flex-1">
-                    <p
-                      className={`${
-                        dark && "text-gray-200"
-                      } text-lg font-semibold leading-8 text-gray-900`}
-                    >
-                      {feature.name}
-                    </p>
-                    <p
-                      className={`${
-                        dark && "text-gray-400"
-                      } mt-2 text-base leading-7 text-gray-600`}
-                    >
-                      {feature.description}
-                    </p>
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-      </div>{" "}
-      {/* ### FEATURES END */}
-      {/* ### TESTIMONIALS START */}
-      <div
-        className={`${dark && "bg-[#030414]"} flex flex-col items-center gap-2`}
-      >
-        <div className="sm:text-center">
-          <h2 className="text-lg font-semibold leading-8 text-blue-600">
-            Testimonials
-          </h2>
-          <p
-            className={`${
-              dark && "text-white"
-            } mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl`}
-          >
-            Customer Reviews
-          </p>
-          <p
-            className={`${
-              dark && "text-gray-400"
-            } mx-auto mt-6 max-w-2xl text-lg leading-8 text-gray-600 mb-5`}
-          >
-            See how Noteslify customers around the globe are becoming productive
-            and managing their time and efficiency.
-          </p>
-        </div>
-        {/*          <!-- Rating --> */}
-        <span className="flex items-center gap-4 px-4 py-2 text-sm rounded shadow-md bg-slate-50 text-slate-500 shadow-slate-200">
-          <span
-            className="flex gap-1 text-amber-400"
-            role="img"
-            aria-label="Rating: 4.2 out of 5 stars"
-          >
-            <span aria-hidden="true">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="currentColor"
-                className="w-6 h-6"
-              >
-                <path
-                  fillRule="evenodd"
-                  d="M10.788 3.21c.448-1.077 1.976-1.077 2.424 0l2.082 5.007 5.404.433c1.164.093 1.636 1.545.749 2.305l-4.117 3.527 1.257 5.273c.271 1.136-.964 2.033-1.96 1.425L12 18.354 7.373 21.18c-.996.608-2.231-.29-1.96-1.425l1.257-5.273-4.117-3.527c-.887-.76-.415-2.212.749-2.305l5.404-.433 2.082-5.006z"
-                  clipRule="evenodd"
-                />
-              </svg>
-            </span>
-            <span aria-hidden="true">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="currentColor"
-                className="w-6 h-6"
-              >
-                <path
-                  fillRule="evenodd"
-                  d="M10.788 3.21c.448-1.077 1.976-1.077 2.424 0l2.082 5.007 5.404.433c1.164.093 1.636 1.545.749 2.305l-4.117 3.527 1.257 5.273c.271 1.136-.964 2.033-1.96 1.425L12 18.354 7.373 21.18c-.996.608-2.231-.29-1.96-1.425l1.257-5.273-4.117-3.527c-.887-.76-.415-2.212.749-2.305l5.404-.433 2.082-5.006z"
-                  clipRule="evenodd"
-                />
-              </svg>
-            </span>
-            <span aria-hidden="true">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="currentColor"
-                className="w-6 h-6"
-              >
-                <path
-                  fillRule="evenodd"
-                  d="M10.788 3.21c.448-1.077 1.976-1.077 2.424 0l2.082 5.007 5.404.433c1.164.093 1.636 1.545.749 2.305l-4.117 3.527 1.257 5.273c.271 1.136-.964 2.033-1.96 1.425L12 18.354 7.373 21.18c-.996.608-2.231-.29-1.96-1.425l1.257-5.273-4.117-3.527c-.887-.76-.415-2.212.749-2.305l5.404-.433 2.082-5.006z"
-                  clipRule="evenodd"
-                />
-              </svg>
-            </span>
-            <span aria-hidden="true">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="currentColor"
-                className="w-6 h-6"
-              >
-                <path
-                  fillRule="evenodd"
-                  d="M10.788 3.21c.448-1.077 1.976-1.077 2.424 0l2.082 5.007 5.404.433c1.164.093 1.636 1.545.749 2.305l-4.117 3.527 1.257 5.273c.271 1.136-.964 2.033-1.96 1.425L12 18.354 7.373 21.18c-.996.608-2.231-.29-1.96-1.425l1.257-5.273-4.117-3.527c-.887-.76-.415-2.212.749-2.305l5.404-.433 2.082-5.006z"
-                  clipRule="evenodd"
-                />
-              </svg>
-            </span>
-            <span aria-hidden="true">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                fill="none"
-                viewBox="0 0 24 24"
-                strokeWidth="1.5"
-                stroke="currentColor"
-                className="w-6 h-6"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  d="M11.48 3.499a.562.562 0 011.04 0l2.125 5.111a.563.563 0 00.475.345l5.518.442c.499.04.701.663.321.988l-4.204 3.602a.563.563 0 00-.182.557l1.285 5.385a.562.562 0 01-.84.61l-4.725-2.885a.563.563 0 00-.586 0L6.982 20.54a.562.562 0 01-.84-.61l1.285-5.386a.562.562 0 00-.182-.557l-4.204-3.602a.563.563 0 01.321-.988l5.518-.442a.563.563 0 00.475-.345L11.48 3.5z"
-                />
-              </svg>
-            </span>
-          </span>
-          <span> 4.2 out 5 </span>
-        </span>
-        {/*          <!-- Helper text --> */}
-        <span className="text-xs leading-6 text-slate-400 mb-10">
-          based on 153 user ratings
-        </span>
-      </div>{" "}
-      {/* ### TESTIMONIALS END */}
+      {/* main */}
+      <main>
+        <MainContainer />
+      </main>
+      {/* features */}
+      <Features dark={dark} />
+      {/* testimonials  */}
+      <Testimonials dark={dark} />
       {/* ### CTA START */}
       <div className={`${dark && "bg-[#131a42]"} bg-gray-50`}>
         <div className="mx-auto max-w-7xl py-12 px-4 sm:px-6 lg:flex lg:items-center lg:justify-between lg:py-16 lg:px-8">
